+++ app/src/main/java/michael/linker/gestureid/data/sensor/listener/composite/GravitySensorListener.java
@@ -45,10 +45,10 @@ public class GravitySensorListener implements ISensorListener {
 
     private void proceedEvent(SensorEvent sensorEvent) {
         GravitySensorModel sensorModel = new GravitySensorModel();
-        sensorModel.setX(sensorEvent.values[0]);
-        sensorModel.setY(sensorEvent.values[1]);
-        sensorModel.setZ(sensorEvent.values[2]);
-        sensorModel.setTimestamp(sensorEvent.timestamp);
+        sensorModel.setX((double) sensorEvent.values[0]);
+        sensorModel.setY((double) sensorEvent.values[1]);
+        sensorModel.setZ((double) sensorEvent.values[2]);
+        sensorModel.setTimestamp((double)sensorEvent.timestamp);
         try {
             eventSynchronizer.registerEvent(sensorModel);
         } catch (EventSynchronizerFailedException e) {
