+++ app/src/main/java/michael/linker/gestureid/data/sensor/listener/composite/LinearAccelerationSensorListener.java
@@ -46,10 +46,10 @@ public class LinearAccelerationSensorListener implements ISensorListener {
 
     private void proceedEvent(SensorEvent sensorEvent) {
         LinearAccelerationSensorModel sensorModel = new LinearAccelerationSensorModel();
-        sensorModel.setX(sensorEvent.values[0]);
-        sensorModel.setY(sensorEvent.values[1]);
-        sensorModel.setZ(sensorEvent.values[2]);
-        sensorModel.setTimestamp(sensorEvent.timestamp);
+        sensorModel.setX((double) sensorEvent.values[0]);
+        sensorModel.setY((double) sensorEvent.values[1]);
+        sensorModel.setZ((double) sensorEvent.values[2]);
+        sensorModel.setTimestamp((double)sensorEvent.timestamp);
         try {
             eventSynchronizer.registerEvent(sensorModel);
         } catch (EventSynchronizerFailedException e) {
