+++ app/src/main/java/michael/linker/gestureid/data/sensor/model/base/AccelerometerSensorModel.java
@@ -3,13 +3,13 @@ package michael.linker.gestureid.data.sensor.model.base;
 import michael.linker.gestureid.data.sensor.model.ThreeAxisSensorModel;
 import michael.linker.gestureid.core.sensor.sensor.type.BaseSensorType;
 
-public final class AccelerometerSensorModel extends ThreeAxisSensorModel<Float> {
+public final class AccelerometerSensorModel extends ThreeAxisSensorModel<Double> {
 
     public AccelerometerSensorModel() {
         super(BaseSensorType.ACCELEROMETER);
     }
 
-    public AccelerometerSensorModel(Long timestamp, Float x, Float y, Float z) {
+    public AccelerometerSensorModel(Double timestamp, Double x, Double y, Double z) {
         super(BaseSensorType.ACCELEROMETER, timestamp, x, y, z);
     }
 }
