+++ app/src/main/java/michael/linker/gestureid/data/sensor/model/base/GyroscopeSensorModel.java
@@ -3,12 +3,12 @@ package michael.linker.gestureid.data.sensor.model.base;
 import michael.linker.gestureid.data.sensor.model.ThreeAxisSensorModel;
 import michael.linker.gestureid.core.sensor.sensor.type.BaseSensorType;
 
-public final class GyroscopeSensorModel extends ThreeAxisSensorModel<Float> {
+public final class GyroscopeSensorModel extends ThreeAxisSensorModel<Double> {
     public GyroscopeSensorModel() {
         super(BaseSensorType.GYROSCOPE);
     }
 
-    public GyroscopeSensorModel(Long timestamp, Float x, Float y, Float z) {
+    public GyroscopeSensorModel(Double timestamp, Double x, Double y, Double z) {
         super(BaseSensorType.GYROSCOPE, timestamp, x, y, z);
     }
 }
