+++ app/src/main/java/michael/linker/gestureid/data/sensor/model/composite/GeomagneticRotationVectorSensorModel.java
@@ -3,12 +3,12 @@ package michael.linker.gestureid.data.sensor.model.composite;
 import michael.linker.gestureid.data.sensor.model.FourAxisSensorModel;
 import michael.linker.gestureid.core.sensor.sensor.type.CompositeSensorType;
 
-public final class GeomagneticRotationVectorSensorModel extends FourAxisSensorModel<Float> {
+public final class GeomagneticRotationVectorSensorModel extends FourAxisSensorModel<Double> {
     public GeomagneticRotationVectorSensorModel() {
         super(CompositeSensorType.GEOMAGNETIC_ROTATION_VECTOR);
     }
 
-    public GeomagneticRotationVectorSensorModel(Long timestamp, Float x, Float y, Float z, Float w) {
+    public GeomagneticRotationVectorSensorModel(Double timestamp, Double x, Double y, Double z, Double w) {
         super(CompositeSensorType.GEOMAGNETIC_ROTATION_VECTOR, timestamp, x, y, z, w);
     }
 }
