+++ app/src/main/java/michael/linker/gestureid/data/sensor/model/composite/GravitySensorModel.java
@@ -3,12 +3,12 @@ package michael.linker.gestureid.data.sensor.model.composite;
 import michael.linker.gestureid.data.sensor.model.ThreeAxisSensorModel;
 import michael.linker.gestureid.core.sensor.sensor.type.CompositeSensorType;
 
-public final class GravitySensorModel extends ThreeAxisSensorModel<Float> {
+public final class GravitySensorModel extends ThreeAxisSensorModel<Double> {
     public GravitySensorModel() {
         super(CompositeSensorType.GRAVITY);
     }
 
-    public GravitySensorModel(Long timestamp, Float x, Float y, Float z) {
+    public GravitySensorModel(Double timestamp, Double x, Double y, Double z) {
         super(CompositeSensorType.GRAVITY, timestamp, x, y, z);
     }
 }
