+++ app/src/main/java/michael/linker/gestureid/data/sensor/model/composite/RotationVectorSensorModel.java
@@ -3,12 +3,12 @@ package michael.linker.gestureid.data.sensor.model.composite;
 import michael.linker.gestureid.data.sensor.model.FourAxisSensorModel;
 import michael.linker.gestureid.core.sensor.sensor.type.CompositeSensorType;
 
-public final class RotationVectorSensorModel extends FourAxisSensorModel<Float> {
+public final class RotationVectorSensorModel extends FourAxisSensorModel<Double> {
     public RotationVectorSensorModel() {
         super(CompositeSensorType.ROTATION_VECTOR);
     }
 
-    public RotationVectorSensorModel(Long timestamp, Float x, Float y, Float z, Float w) {
+    public RotationVectorSensorModel(Double timestamp, Double x, Double y, Double z, Double w) {
         super(CompositeSensorType.ROTATION_VECTOR, timestamp, x, y, z, w);
     }
 }
